Not a happy state for the way the controls are architected - has some serious smell but it's been refactored to work at the very least which is what matters. can start doing some better refactoring work here
