+++ package.json
@@ -17,6 +17,7 @@
         "react-router": "^7.0.2",
         "react-toggle": "^4.1.3",
         "sass": "^1.69.5",
+        "three": "^0.171.0",
         "uuid": "^11.0.3",
         "zustand": "^5.0.2"
     },
@@ -49,12 +50,13 @@
     "devDependencies": {
         "@eslint/js": "^9.17.0",
         "@react-three/eslint-plugin": "^0.1.1",
+        "@types/node": "^22.10.2",
         "@types/react": "18.3.1",
         "@types/react-dom": "18.3.1",
         "@types/react-toggle": "^4.0.5",
-        "@types/uuid": " ^10.0.0",
-        "@types/node": "^22.10.2",
         "@types/three": "^0.171.0",
+        "@types/uuid": " ^10.0.0",
+        "@vitejs/plugin-react-swc": "^3.4.1",
         "eslint": "^9.17.0",
         "eslint-config-stylelint": "^23.0.0",
         "eslint-import-resolver-typescript": "^3.7.0",
@@ -71,7 +73,6 @@
         "typescript-eslint": "^8.18.2",
         "vite": "^6.0.3",
         "vite-plugin-svgr": "^4.1.0",
-        "vite-tsconfig-paths": "^5.1.4",
-        "@vitejs/plugin-react-swc": "^3.4.1"
+        "vite-tsconfig-paths": "^5.1.4"
     }
 }
