+++ src/app/Pages/Home/components/Content/Content.module.scss
@@ -5,5 +5,4 @@
 	width: 100vw;
 	position: sticky;
 	transition: background-color 350ms;
-
 }
