+++ src/app/Pages/Home/components/Content/components/InteractiveIsland/components/Island/Lamps.tsx
