+++ src/app/Pages/Home/components/Content/components/InteractiveIsland/components/Lights/Lights.tsx
