+++ src/app/Pages/Home/components/Content/components/InteractiveIsland/components/Zotbot/WheelJoint.tsx
